requirements.txt
ansible==2.4.1.0
Django==2.0.5
django-celery-beat==1.2.0
django-celery-results==1.0.1
django-cors-headers==2.5.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.0.1
django-reversion==2.0.13
django-timezone-field==3.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.15.2
gevent==1.3.7
gitdb2==2.0.5
GitPython==2.1.11
greenlet==0.4.15
hiredis==0.2.0
html5lib==1.0.1
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jdcal==1.4
jieba==0.39
Jinja2==2.10
kombu==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
msgpack-python==0.5.6
nose==1.3.7
odfpy==1.3.6
openpyxl==2.5.3
paramiko==2.4.1
Pillow==5.2.0
ply==3.11
pyasn1==0.4.2
pycparser==2.18
pycryptodomex==3.6.1
pyguacamole==0.8
pymongo==3.6.1
PyMySQL==0.8.1
PyNaCl==1.2.1
pysmi==0.3.0
pysnmp==4.4.4
python-crontab==2.3.5
python-dateutil==2.7.3
python-nmap==0.6.1
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.19.1
six==1.10.0
smmap2==2.0.5
svn==0.3.46
tablib==0.12.1
Twisted==17.5.0
txaio==2.10.0
unicodecsv==0.14.1
urllib3==1.23
vine==1.1.4
webencodings==0.5.1
websockets==5.0.1
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
pyjwt
pymysql
